More refactoring and commenting, for code clarity
