+++ src/main/resources/templates/fragments/push.html
@@ -4,9 +4,13 @@
 
 <div data-th-fragment="fragment-push" data-th-remove="tag">
     <div id="page-content">
-        <form class="push-forms subscribe" data-th-action="${subscribeUrl}" onSubmit="return false;">
+        <form class="push-forms subscribe"
+              id="subscribe-form"
+              data-th-action="${subscribeUrl}"
+              th:attr="data-public-key=${publicKey}"
+              onSubmit="return false;">
             <label id="subscribe-status">Subscribe to notifications</label>
-            <div class="button-holder"><button id="subscribe-button">Yes please</button></div>
+            <div class="button-holder"><button id="subscribe-button">Subscribe</button></div>
         </form>
 
         <form class="push-forms push disabled" id="push-form" data-th-action="${pushUrl}" onSubmit="return false;">
@@ -17,10 +21,6 @@
     </div>
 
     <div data-th-replace="/templates/fragments/common::fragment-no-javascript"></div>
-
-    <div id="push-subscription-data" style="display:none;"
-         th:attr="data-subscribe-url=${subscribeUrl},data-public-key=${publicKey}">
-    </div>
 
     <script defer data-th-src="${portal.assetUrl({'_path=js/push.js'})}" type="text/javascript"></script>
 </div>
